requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
billiard==3.3.0.22
boto==2.39.0
celery==3.1.20
coverage==4.0.3
decorator==4.0.9
dj-database-url==0.4.0
Django==1.9.2
django-filter==0.12.0
django-naomi==0.8
django-storages==1.4
djangorestframework==3.3.2
gnureadline==6.3.3
gunicorn==19.4.5
ipdb==0.8.1
ipython==4.1.1
ipython-genutils==0.1.0
kombu==3.0.33
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
python-decouple==3.0
pytz==2015.7
simplegeneric==0.8.1
traitlets==4.1.0
whitenoise==2.0.6
django-authtools==1.4.0
readability-api==1.0.2
